Add a build flag for the undefined behavior sanitizer, aka "ubsan".
Ubsan checks for undefined behavior according to the C++
standard. Some of this behavior has been known to be exploited by
optimizing compilers to produce bizarre results, like taking both
branches of a conditional.

Although a number of cases of UB were not possible to fix, as they
were caused by libraries like Thrift or libstdc++, this patch fixes
the UB that was feasible to fix.

Change-Id: I88c7234bd7c5eb7404490a0913d90470c10835e7